backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
google-auth==2.35.0
pydantic==2.9.2
python-multipart==0.0.12
bcrypt==4.2.0